requirements.txt
grpcio>=1.80.0
grpcio-tools>=1.80.0
pika>=1.3.0
protobuf>=6.31.0
psycopg2-binary>=2.9.0